+++ spec/classes/storm/storm_gridftp_install_spec.rb
@@ -1,11 +1,8 @@
 require 'spec_helper'
 
 describe 'storm::gridftp::install' do
-
   on_supported_os.each do |os, facts|
-
     context "on #{os}" do
-
       let(:facts) do
         facts
       end
@@ -13,7 +10,6 @@ describe 'storm::gridftp::install' do
       it 'check storm globus gridftp server is installed' do
         is_expected.to contain_package('storm-globus-gridftp-mp').with(ensure: 'installed')
       end
-
     end
   end
 end