+++ spec/classes/storm/storm_gridftp_service_spec.rb
@@ -1,12 +1,11 @@
 require 'spec_helper'
 
 describe 'storm::gridftp::service' do
-
   on_supported_os.each do |os, facts|
-
     context "on #{os}" do
-
-      let(:pre_condition) { 'include storm::gridftp::install' }
+      let(:pre_condition) do
+        'include storm::gridftp::install'
+      end
 
       let(:facts) do
         facts
@@ -14,8 +13,6 @@ describe 'storm::gridftp::service' do
 
       it { is_expected.to compile.with_all_deps }
       it { is_expected.to contain_service('storm-globus-gridftp').with(ensure: 'running') }
-
     end
   end
-
 end