+++ spec/defines/storm_webdav_storage_area_file_spec.rb
@@ -1,11 +1,8 @@
 require 'spec_helper'
 
 describe 'storm::webdav::storage_area_file' do
-
   on_supported_os.each do |os, facts|
-
     context "on #{os}" do
-
       let(:pre_condition) do
         <<-EOF
           class { 'storm::webdav':
@@ -13,7 +10,7 @@ describe 'storm::webdav::storage_area_file' do
           }
         EOF
       end
-      
+
       let(:facts) do
         facts
       end
@@ -26,13 +23,12 @@ describe 'storm::webdav::storage_area_file' do
         }
       end
 
-      it "check storage area file exists" do
-        storage_area_file='/etc/storm/webdav/sa.d/test.vo.properties'
+      it 'check storage area file exists' do
+        storage_area_file = '/etc/storm/webdav/sa.d/test.vo.properties'
         is_expected.to contain_file(storage_area_file).with(
-          :source => '/path/to/test.vo.properties',
+          source: '/path/to/test.vo.properties',
         )
       end
-
     end
   end
 end